3.2.0-beta

+++ Linux/install.sh
@@ -11,38 +11,59 @@ echo "**************************************************************************
 set -e
 
 folder=$(pwd);
-
+echo $folder;
 id=$USER;
 
-#stop error detection
-echo "Checking if Scrot is installed..!!";
+echo "Checking Scrot Installation";
 set +e
-if [ $(dpkg-query -W -f='${Status}' scrot 2>/dev/null | grep -c "ok installed") -eq 0 ];
-then
-  echo "Installing scrot";
-  sudo apt-get install scrot;
-fi
+
+if [ -f /etc/redhat-release]; then
+    echo "Installing Scrot For RedHat: \n"
+
+      if yum list installed scrot >/dev/null 2>&1; then
+        echo "Scrot Is Already Installed, Skipping Installation."
+      else
+          sudo yum install scrot;
+      fi
+
+elif [ -f /etc/debian_version]; then
+    if [ $(dpkg-query -W -f='${Status}' scrot 2>/dev/null | grep -c "ok installed") -eq 0 ];
+    then
+      echo "Installing Scrot For Debian/Ubuntu: \n";
+      sudo apt-get install scrot;
+    fi
+
+else
+    echo "Installing Scrot: \n";
+    cd scrot;
+    ./configure;
+    make;
+    su -c "make install";
+    cd ../;
+
+
+
 
 #start error detection
 set -e
 
-mkdir ~/Desktop/Myss;
-mkdir ~/Desktop/Myss/scripts;
-mkdir ~/Desktop/Myss/ss;
+mkdir -p ~/Desktop/Myss;
+mkdir -p ~/Desktop/Myss/scripts;
+mkdir -p ~/Desktop/Myss/ss;
 echo "Folder created on desktop for stroing screenshots";
 
-sudo chmod 755 ~/$folder/myss-master/Linux/Myss.desktop;
-sudo chmod 755 ~/$folder/myss-master/Linux/compress-myss.desktop;
-sudo chmod 755 ~/$folder/myss-master/Linux/myss.sh;
-sudo chmod 755 ~/$folder/myss-master/Linux/compress-myss.sh
+sudo chmod 755 $folder/Myss.desktop;
+sudo chmod 755 $folder/compress-myss.desktop;
+sudo chmod 755 $folder/myss.sh;
+sudo chmod 755 $folder/compress-myss.sh;
 
-cp ~/$folder/myss-master/Linux/myss.sh ~/Desktop/Myss/scripts/myss.sh ;
-cp ~/$folder/myss-master/Linux/compress-myss.sh ~/Desktop/Myss/scripts/compress-myss.sh;
-cp ~/$folder/myss-master/Linux/Myss.desktop ~/Desktop/Myss/;
-cp ~/$folder/myss-master/Linux/compress-myss.desktop ~/Desktop/Myss/;
+cp $folder/myss.sh ~/Desktop/Myss/scripts/myss.sh ;
+cp $folder/compress-myss.sh ~/Desktop/Myss/scripts/compress-myss.sh;
+cp $folder/Myss.desktop ~/Desktop/Myss/;
+cp $folder/compress-myss.desktop ~/Desktop/Myss/;
 
-echo Exec=gnome-terminal -e /home/$id/Desktop/Myss/scripts/myss.sh >> ~/Desktop/Myss/Myss.desktop;
-echo Exec=gnome-terminal -e /home/$id/Desktop/Myss/scripts/compress-myss.sh >> ~/Desktop/Myss/compress-myss.desktop;
+echo Exec=gnome-terminal -e ~/Desktop/Myss/scripts/myss.sh >> ~/Desktop/Myss/Myss.desktop;
+echo Exec=gnome-terminal -e ~/Desktop/Myss/scripts/compress-myss.sh >> ~/Desktop/Myss/compress-myss.desktop;
 
 echo "Sucess!";
 echo "Myss has been installed, to start taking screenhsots run the myss file from desktop.";
